requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.18
pdf2docx>=0.5.8
Pillow>=11.1.0
python-docx>=1.1.2
PyMuPDF>=1.25.0
pytesseract>=0.3.10
